requirements.txt
unidecode
click
coloredlogs
networkx
chardet
python-magic
python-magic-bin; sys_platform=="win32"